requirements.txt
jsonschema>=4.0.0
pytest>=7.0.0
ansible-lint>=6.0.0
requests>=2.28.0
PyYAML>=6.0
